standalone: split standalone into core setup artifact (as needed for LSP) and a real standalone, usable as standalone application with pre-configured piccolo exporter dependency. Restores old behavior of standalone in last release.
