--- plugins/de.cau.cs.kieler.klighd.lsp/META-INF/MANIFEST.MF
@@ -1,39 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: KIELER Lightweight Diagrams as Language Server
-Bundle-SymbolicName: de.cau.cs.kieler.klighd.lsp;singleton:=true
-Bundle-Version: 2.3.1.qualifier
-Bundle-ClassPath: .
-Bundle-RequiredExecutionEnvironment: JavaSE-11
-Export-Package: de.cau.cs.kieler.klighd.lsp,
- de.cau.cs.kieler.klighd.lsp.gson_utils,
- de.cau.cs.kieler.klighd.lsp.interactive,
- de.cau.cs.kieler.klighd.lsp.interactive.layered,
- de.cau.cs.kieler.klighd.lsp.interactive.mrtree,
- de.cau.cs.kieler.klighd.lsp.interactive.rectpacking,
- de.cau.cs.kieler.klighd.lsp.launch,
- de.cau.cs.kieler.klighd.lsp.model,
- de.cau.cs.kieler.klighd.lsp.utils
-Bundle-Vendor: Kiel University
-Bundle-Localization: plugin
-Bundle-ActivationPolicy: lazy
-Require-Bundle: de.cau.cs.kieler.kgraph.text,
- de.cau.cs.kieler.klighd,
- de.cau.cs.kieler.klighd.ide,
- de.cau.cs.kieler.klighd.kgraph,
- de.cau.cs.kieler.klighd.standalone,
- com.google.gson;bundle-version="2.8.2",
- com.google.inject;bundle-version="3.0.0",
- org.apache.log4j;bundle-version="1.2.15",
- org.eclipse.elk.alg.layered,
- org.eclipse.elk.alg.mrtree,
- org.eclipse.elk.alg.rectpacking,
- org.eclipse.elk.core,
- org.eclipse.elk.graph,
- org.eclipse.emf.ecore;bundle-version="2.14.0",
- org.eclipse.lsp4j;bundle-version="0.8.0",
- org.eclipse.lsp4j.generator;bundle-version="0.8.0",
- org.eclipse.lsp4j.jsonrpc;bundle-version="0.8.0",
- org.eclipse.sprotty;bundle-version="0.7.0",
- org.eclipse.xtext.ide;bundle-version="2.21.0",
- org.eclipse.xtext.xbase;bundle-version="2.21.0"
