+++ plugins/de.cau.cs.kieler.klighd.lsp/pom.xml
@@ -69,7 +69,7 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>de.cau.cs.kieler.klighd.standalone</artifactId>
+      <artifactId>de.cau.cs.kieler.klighd.setup</artifactId>
       <version>${project.version}</version>
       <type>jar</type>
     </dependency>
