+++ plugins/de.cau.cs.kieler.klighd.piccolo.freehep/.classpath
@@ -1,13 +1,27 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-11"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER">
+		<attributes>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
-	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="src" output="bin" path="src">
+		<attributes>
+			<attribute name="optional" value="true"/>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry kind="lib" path="lib/freehep-graphics2d-2.3.jar" sourcepath="lib/src/freehep-graphics2d-2.3-sources.jar"/>
 	<classpathentry kind="lib" path="lib/freehep-graphicsbase-2.3.jar" sourcepath="lib/src/freehep-graphicsbase-2.3-sources.jar"/>
 	<classpathentry kind="lib" path="lib/freehep-graphicsio-2.3.jar" sourcepath="lib/src/freehep-graphicsio-2.3-sources.jar"/>
 	<classpathentry kind="lib" path="lib/freehep-graphicsio-svg-2.3.jar" sourcepath="lib/src/freehep-graphicsio-svg-2.3-sources.jar"/>
 	<classpathentry kind="lib" path="lib/freehep-graphicsio-tests-2.3.jar" sourcepath="lib/src/freehep-graphicsio-tests-2.3-sources.jar"/>
 	<classpathentry kind="lib" path="lib/freehep-io-2.2.2.jar" sourcepath="lib/src/freehep-io-2.2.2-sources.jar"/>
+	<classpathentry kind="con" path="org.eclipse.m2e.MAVEN2_CLASSPATH_CONTAINER">
+		<attributes>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
