+++ plugins/de.cau.cs.kieler.klighd.setup/.settings/org.eclipse.jdt.ui.prefs
