+++ plugins/de.cau.cs.kieler.klighd.setup/.settings/org.eclipse.xtend.core.Xtend.prefs
@@ -2,7 +2,5 @@ BuilderConfiguration.is_project_specific=true
 eclipse.preferences.version=1
 outlet.DEFAULT_OUTPUT.hideLocalSyntheticVariables=true
 outlet.DEFAULT_OUTPUT.installDslAsPrimarySource=false
-outlet.DEFAULT_OUTPUT.sourceFolder.src-gen.directory=xtend-gen
 outlet.DEFAULT_OUTPUT.sourceFolder.src.directory=xtend-gen
-outlet.DEFAULT_OUTPUT.sourceFolder.xtend-gen.directory=xtend-gen
 outlet.DEFAULT_OUTPUT.userOutputPerSourceFolder=true
