+++ plugins/de.cau.cs.kieler.klighd.setup/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+bin.includes = META-INF/,\
+               .,\
+               resources/
