+++ plugins/de.cau.cs.kieler.klighd.setup/pom.xml
@@ -1,0 +1,133 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>de.cau.cs.kieler.klighd</groupId>
+    <artifactId>parent</artifactId>
+    <version>2.3.1-SNAPSHOT</version>
+    <relativePath>../../pom.xml</relativePath>
+  </parent>
+
+  <artifactId>de.cau.cs.kieler.klighd.setup</artifactId>
+  <packaging>eclipse-plugin</packaging>
+
+  <name>KIELER Lightweight Diagrams Standalone Setup</name>
+  <description>KIELER Lightweight Diagrams Standalone Setup</description>
+  <url>https://github.com/kieler/KLighD</url>
+
+  <dependencies>
+    <dependency>
+      <groupId>com.google.inject</groupId>
+      <artifactId>guice</artifactId>
+      <version>${guice-version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.google.guava</groupId>
+      <artifactId>guava</artifactId>
+      <version>${guava-version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.elk</groupId>
+      <artifactId>org.eclipse.elk.core.service</artifactId>
+      <version>${elk-version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.elk</groupId>
+      <artifactId>org.eclipse.elk.alg.force</artifactId>
+      <version>${elk-version}</version>
+      <scope>compile</scope>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.elk</groupId>
+      <artifactId>org.eclipse.elk.alg.graphviz.layouter</artifactId>
+      <version>${elk-version}</version>
+      <scope>compile</scope>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.elk</groupId>
+      <artifactId>org.eclipse.elk.alg.layered</artifactId>
+      <version>${elk-version}</version>
+      <scope>compile</scope>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.elk</groupId>
+      <artifactId>org.eclipse.elk.alg.mrtree</artifactId>
+      <version>${elk-version}</version>
+      <scope>compile</scope>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>org.eclipse.elk</groupId>
+      <artifactId>org.eclipse.elk.alg.radial</artifactId>
+      <version>${elk-version}</version>
+      <scope>compile</scope>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>de.cau.cs.kieler.klighd</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>de.cau.cs.kieler.klighd.kgraph</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>de.cau.cs.kieler.klighd.krendering</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>de.cau.cs.kieler.klighd.krendering.extensions</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+  </dependencies>
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.eclipse.elk</groupId>
+        <artifactId>org.eclipse.elk.alg.force</artifactId>
+        <version>${elk-version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.elk</groupId>
+        <artifactId>org.eclipse.elk.alg.graphviz.dot</artifactId>
+        <version>${elk-version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.elk</groupId>
+        <artifactId>org.eclipse.elk.alg.graphviz.layouter</artifactId>
+        <version>${elk-version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.elk</groupId>
+        <artifactId>org.eclipse.elk.alg.layered</artifactId>
+        <version>${elk-version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.elk</groupId>
+        <artifactId>org.eclipse.elk.alg.mrtree</artifactId>
+        <version>${elk-version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.elk</groupId>
+        <artifactId>org.eclipse.elk.alg.radial</artifactId>
+        <version>${elk-version}</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+
+  <build>
+    <resources>
+     <resource>
+       <directory>resources</directory>
+     </resource>
+   </resources>
+  </build>
+</project>
