+++ plugins/de.cau.cs.kieler.klighd.setup/resources/fonts/LICENSE_FONTS.txt
