+++ plugins/de.cau.cs.kieler.klighd.standalone/build.properties
@@ -1,4 +1,2 @@
-source.. = src/
 bin.includes = META-INF/,\
-               .,\
-               resources/
+               .
