+++ plugins/de.cau.cs.kieler.klighd.standalone/pom.xml
@@ -9,127 +9,28 @@
     <relativePath>../../pom.xml</relativePath>
   </parent>
 
-  <groupId>de.cau.cs.kieler.klighd</groupId>
   <artifactId>de.cau.cs.kieler.klighd.standalone</artifactId>
-  <version>2.3.1-SNAPSHOT</version>
   <packaging>eclipse-plugin</packaging>
 
-  <name>KIELER Lightweight Diagrams Standalone Setup</name>
-  <description>KIELER Lightweight Diagrams Standalone Setup</description>
+  <name>KIELER Lightweight Diagrams Standalone</name>
+  <description>KIELER Lightweight Diagrams Standalone</description>
   <url>https://github.com/kieler/KLighD</url>
 
   <dependencies>
     <dependency>
-      <groupId>com.google.inject</groupId>
-      <artifactId>guice</artifactId>
-      <version>${guice-version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-      <version>${guava-version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.elk</groupId>
-      <artifactId>org.eclipse.elk.core.service</artifactId>
-      <version>${elk-version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.elk</groupId>
-      <artifactId>org.eclipse.elk.alg.force</artifactId>
-      <version>${elk-version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.elk</groupId>
-      <artifactId>org.eclipse.elk.alg.graphviz.layouter</artifactId>
-      <version>${elk-version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.elk</groupId>
-      <artifactId>org.eclipse.elk.alg.layered</artifactId>
-      <version>${elk-version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.elk</groupId>
-      <artifactId>org.eclipse.elk.alg.mrtree</artifactId>
-      <version>${elk-version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.elk</groupId>
-      <artifactId>org.eclipse.elk.alg.radial</artifactId>
-      <version>${elk-version}</version>
-      <scope>compile</scope>
-      <optional>true</optional>
-    </dependency>
-    <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>de.cau.cs.kieler.klighd</artifactId>
+      <artifactId>de.cau.cs.kieler.klighd.setup</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>de.cau.cs.kieler.klighd.kgraph</artifactId>
+      <artifactId>de.cau.cs.kieler.klighd.piccolo</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>de.cau.cs.kieler.klighd.krendering</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>${project.groupId}</groupId>
-      <artifactId>de.cau.cs.kieler.klighd.krendering.extensions</artifactId>
+      <artifactId>de.cau.cs.kieler.klighd.piccolo.freehep</artifactId>
       <version>${project.version}</version>
     </dependency>
   </dependencies>
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.eclipse.elk</groupId>
-        <artifactId>org.eclipse.elk.alg.force</artifactId>
-        <version>${elk-version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse.elk</groupId>
-        <artifactId>org.eclipse.elk.alg.graphviz.dot</artifactId>
-        <version>${elk-version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse.elk</groupId>
-        <artifactId>org.eclipse.elk.alg.graphviz.layouter</artifactId>
-        <version>${elk-version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse.elk</groupId>
-        <artifactId>org.eclipse.elk.alg.layered</artifactId>
-        <version>${elk-version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse.elk</groupId>
-        <artifactId>org.eclipse.elk.alg.mrtree</artifactId>
-        <version>${elk-version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse.elk</groupId>
-        <artifactId>org.eclipse.elk.alg.radial</artifactId>
-        <version>${elk-version}</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <build>
-    <resources>
-     <resource>
-       <directory>resources</directory>
-     </resource>
-   </resources>
-  </build>
 </project>
