+++ pom.xml
@@ -23,6 +23,7 @@
     <module>plugins/de.cau.cs.kieler.klighd.krendering.extensions</module>
     <module>plugins/de.cau.cs.kieler.klighd.lsp</module>
     <module>plugins/de.cau.cs.kieler.klighd.piccolo</module>
+    <module>plugins/de.cau.cs.kieler.klighd.setup</module>
     <module>plugins/de.cau.cs.kieler.klighd.standalone</module>
     <module>test/de.cau.cs.kieler.klighd.test</module>
     <module>test/de.cau.cs.kieler.klighd.piccolo.test</module>
